Reorganize and expand test suite layout
- Relocate correctness/oracle/property test modules into dedicated
  tests/ subpackages under _etrace_algorithms and _etrace_compiler.
- Split _legacy/_legacy_test.py into _ops_test.py and _params_test.py,
  mirroring the _ops.py / _params.py module boundary.
- Merge api_contract_test.py and legacy_deprecation_test.py into
  __init___test.py, the test file corresponding to the package root
  braintrace/__init__.py.
